Raise CannotConnect instead of ConnectionError in validation method
