+++ tests/components/ws66i/test_init.py
@@ -2,13 +2,16 @@
 from unittest.mock import patch
 
 from homeassistant.components.ws66i.const import DOMAIN
-from homeassistant.const import CONF_IP_ADDRESS
+from homeassistant.config_entries import ConfigEntryState
 
-from .test_media_player import MOCK_CONFIG, MOCK_DEFAULT_OPTIONS, MockWs66i
+from .test_media_player import (
+    MOCK_CONFIG,
+    MOCK_DEFAULT_OPTIONS,
+    MOCK_OPTIONS,
+    MockWs66i,
+)
 
 from tests.common import MockConfigEntry
-
-CONFIG = {CONF_IP_ADDRESS: "1.1.1.1"}
 
 ZONE_1_ID = "media_player.zone_11"
 
@@ -19,10 +22,14 @@ async def test_cannot_connect(hass):
         "homeassistant.components.ws66i.get_ws66i",
         new=lambda *a: MockWs66i(fail_open=True),
     ):
-        config_entry = MockConfigEntry(domain=DOMAIN, data=MOCK_CONFIG)
+        config_entry = MockConfigEntry(
+            domain=DOMAIN, data=MOCK_CONFIG, options=MOCK_OPTIONS
+        )
         config_entry.add_to_hass(hass)
         await hass.config_entries.async_setup(config_entry.entry_id)
         await hass.async_block_till_done()
+
+        assert config_entry.state is ConfigEntryState.SETUP_RETRY
         assert hass.states.get(ZONE_1_ID) is None
 
 
@@ -43,4 +50,29 @@ async def test_cannot_connect_2(hass):
             await hass.config_entries.async_setup(config_entry.entry_id)
             await hass.async_block_till_done()
 
+        assert config_entry.state is ConfigEntryState.SETUP_RETRY
         assert hass.states.get(ZONE_1_ID) is None
+
+
+async def test_unload_config_entry(hass):
+    """Test unloading config entry."""
+    with patch(
+        "homeassistant.components.ws66i.get_ws66i",
+        new=lambda *a: MockWs66i(),
+    ):
+        config_entry = MockConfigEntry(
+            domain=DOMAIN, data=MOCK_CONFIG, options=MOCK_OPTIONS
+        )
+        config_entry.add_to_hass(hass)
+        await hass.config_entries.async_setup(config_entry.entry_id)
+        await hass.async_block_till_done()
+
+        assert hass.data[DOMAIN][config_entry.entry_id]
+
+    with patch.object(MockWs66i, "close") as method_call:
+        await config_entry.async_unload(hass)
+        await hass.async_block_till_done()
+
+        assert method_call.called
+
+    assert not hass.data[DOMAIN]
